fix: handle commands even when an update is running
Previously control commands were ignored/delayed while the update
command was running.
Now the update command is put into a background task, so commands can
still be processed. This is important for eInk displays, as the update
can take very long (tens of seconds).
Currently if any new update is scheduled while an old one is still
running, the update is ignored.