feat: buffer image updates that are scheduled during an update
If a command for an image update arrives during a running image update,
then this command will now be buffered to be executed once the update
has finished. Previously it was discarded. Updates are now only
discarded if a third update arrives while the first one is still
executing. Then the third update will replace the buffered second
update.