feat: expand leading ~ in gallery folder list
If a gallery contains paths that start with a `~`, the tilde is now
expanded to the users home directory.
Tilde expansion is only performed for the first path element, and only
the current user is supported. Referencing the homes of other users
(e.g. ~john/some/path) is not supported.