feat: respond to incoming requests
The daemon now sends a response back to the requester once a request has
been handled. Currently this only includes error information or not.
The request is always sent after an update has finished, this may be
changed in the future (e.g. with a specific wait for update request).
The MQTT listener does not yet publish responses back on the broker.

Items: WK-5