feat: send replies to MQTT requests
Updated the MQTT listener to send responses, if a key "response_topic"
is given in the request. Optionally correlation data can be included
with the request, which will be forwarded to the response.
Furthermore a new "BadRequest" response was added, which is sent if a
request could not be parsed. Before this change, the server would exit
on a bad message, now it is handled gracefully.

Due to a quirk in the MQTT implementation, BadRequests sent via MQTT do
not result in a "BadRequest" response, but are instead ignored.

Items: WK-7