feat: avoid selecting recently selected images
Added a reroll system to avoid selecting the same image twice in a row.
The daemon now keeps a list of recently selected images. If a newly
selected image is in that list, then it will instead randomly select a
new one, up to a configurable amount of retries.

Item: Wk-11