fix: correctly start if old socket files exist
The gallerica daemon would sometimes fail to start, because of old unix
socket files. For example if the daemon is killed, it leaves the domain
socket file behind. On startup it would then refuse to start (Address
already in use).

If such files are encountered during startup now, then the daemon will
test if it is still in use or not. If the file is unused, it will be
replaced by a new domain socket and the daemon will start correctly.
If the socket is used, the daemon will still abort with an "Address
already in use" error.